CI unit tests GHA workflow (#80)
Adds a GHA workflow for building and running `fvdb-reality-capture`
pytests in CI. This requires fvdb, so for now it just clones
`fvdb-core`, builds it, uploads a wheel, and then on a GPU runner it
downloads that wheel, builds the fvdb-reality-capture wheel, and runs
pytests.

This currently uses fvdb conda environments. Most of this is copied from
the fvdb CI workflow of the same name, modified as needed.

---------

Signed-off-by: Mark Harris <mharris@nvidia.com>