Mega PR for GTC DC (#96)
Apologies to those who have to review this.

I wrote a notebook to walk a user through how to load a 3d capture,
transform it, reconstruct a gaussian splat scene, visualize the scene
and reconstruct a mesh. Along the way, I realized that the API for our
runner was written in a way that made it useful for the CLI but not very
useful for writing your own scripts. For example, the runner had a fixed
prescribed set of scene transforms and always relied on loading a scene
from the file system. If you're hacking around with scene transforms,
you probably want to construct a viewer directly from a scene. Also, the
name SceneOptmizationRunner is uninformative. How is it optimizing a
scene? I renamed this class to GaussianSplatReconstruction, following
the convention in scikit-learn where engines are named after the model.
I also managed to get rid of the checkpoint class entirely since we're
now serializing SfmScenes and instead the GaussianSplatReconstruct class
just saves a dict.

I also promoted usdz export to its own tool, and made a script in the
main repo which converts a ply to a USDZ.

To showcase all of this, there is now a notebook called
reconstruct_gaussian_splats.ipynb which gives a nice tour of our
features.

---------

Signed-off-by: Francis Williams <francis@fwilliams.info>