Defer logging to avoid extra synchronization (#144)
`.item()` triggers host-device stream synchronization for both single
and multi-GPU since it needs to copy the contents of the tensor element
into a CPU scalar. Calling `loss.item()` twice within the loop at
separate points reduces the throughput since we need to wait for the
kernel streams to flush twice instead of once. Instead, we defer the
first call in the pbar update until later on in the loop where we log
the loss anyhow.

This increases dual GPU speeds by about 10% and single GPU speeds about
5-7%.

Signed-off-by: Matthew Cong <mcong@nvidia.com>