Update comparison benchmark for new fvdb reality capture API (#140)
Fixes #74 

This fixes fvdb and gsplat comparisons in the comparison benchmark by
updating to the new fvdb-reality-capture API and re-enabling support for
running gsplat.

This also readds the --plot-only command so that plots can be modified
without rerunning the benchmarks. Fixes up the plots and adds control
over colors to the opt-config files.

Finally, since the containers now work for both Blackwell and also
earlier GPUs with the same version of CUDA and PyTorch, this PR
eliminates the docker configuration for Blackwell, which greatly
simplifies the docker setup and documentation.

---------

Signed-off-by: Mark Harris <mharris@nvidia.com>